fix: resolve mypy errors and plot_slippage_sensitivity test failure
- timing.py: cast return value to float for market_timing_return()
- bokeh_backend.py: add type: ignore for figure() kwargs (incomplete stubs)
- transactions.py: use ax.plot() instead of Series.plot() to avoid pandas
  date formatting on integer index (fixes TypeError: index type not supported)

Made-with: Cursor

+++ fincore/metrics/timing.py
@@ -174,7 +174,7 @@ def market_timing_return(
 
     excess_factor = factor_aligned - risk_free
 
-    return gamma * np.mean(excess_factor**2)
+    return float(gamma * np.mean(excess_factor**2))
 
 
 def cornell_timing(
